% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchgeo.R
\docType{methods}
\name{fetchGeoFile}
\alias{fetchGeoFile}
\alias{fetchGeoFile,CrunchGeography-method}
\alias{fetchGeoFile,CrunchVariable-method}
\alias{fetchGeoFile,ANY-method}
\title{Fetch a geo- or topo-json hosted by Crunch}
\usage{
fetchGeoFile(geography, ...)

\S4method{fetchGeoFile}{CrunchGeography}(geography, ...)

\S4method{fetchGeoFile}{CrunchVariable}(geography, ...)

\S4method{fetchGeoFile}{ANY}(geography, ...)
}
\arguments{
\item{geography}{a \code{CrunchGeography} object, generally obtained by using the \code{geo}
function on a dataset variable that has geography data specified.}

\item{...}{additional options to pass to \code{geojson_read} or \code{topojson_read}}
}
\description{
Fetch a geo- or topo-json hosted by Crunch
}
\examples{
\dontrun{
geo_data <- fetchGeoFile(geo(ds$country))
}

}
