% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunchgeo.R
\name{getGeoDataFrame}
\alias{getGeoDataFrame}
\title{Get a CrunchDataFrame inside of a SpatialDataFrame}
\usage{
getGeoDataFrame(geo_var, data, ...)
}
\arguments{
\item{geo_var}{the alias of the variable that has geodata associated with it}

\item{data}{a Crunch dataset object}

\item{...}{passed to \code{fetchGeoFile}}
}
\value{
a SpatialDataFrame with a CrunchDataFrame in the \code{Data} slot instead
of a standard data.frame
}
\description{
Get a CrunchDataFrame inside of a SpatialDataFrame
}
