% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-helpers.R
\name{leaflet_hover}
\alias{leaflet_hover}
\title{Helper for creating leaflet labels on mouse over}
\usage{
leaflet_hover(title, item, value)
}
\arguments{
\item{title}{the title of the hover over}

\item{item}{the item that is bein displayed?}

\item{value}{the value to be displayed (currently ends with a percent sign)}
}
\value{
list of html widgets, one for each title, item, value.
}
\description{
This function creates the html necesary for making labels in leaflet. It is
vectorized so you can provide it with vectors of titles, items, and values
and it will make one label for each (so that each region on a map has it's
own hover over, for example)
}
